refactor(frontend): split GameProvider, drop dead raid route (P0-2)
Removes a parallel route-based raid implementation that used message
types the backend never sends (raid_started, boss_damaged). Drops
/raid and /raid/:roomCode routes; the single raid entry point is the
in-game RaidView. GameProvider shrinks 385->157 lines (16->9 useStates)
by extracting useEndlessSettings, usePlayerHealth, usePotionSystem,
and useMonsterAttackLoop sub-hooks.
